+++ src/main/resources/application.yaml
@@ -22,6 +22,10 @@ integrations:
     tokenLocation: file:/var/run/secrets/kubernetes.io/serviceaccount/token
     url: https://kubernetes.default.svc.cluster.local
 
+gobo:
+    wait:
+        afterDeletePodTime: 500
+
 ---
 spring:
     profiles: local
